refactor [#69] Token과 cookie 관련 util 파일 분리
